test if is file before deletion

+++ contest/routines.py
@@ -474,7 +474,9 @@ def handle_uploaded_file(atempt, f, contest):
 	atempt.grade = (round(pct / 100 * contest.max_classification, 0), 0)[mandatory_failed]
 	atempt.save()
 
-	os.remove(os.path.join(paths['dir'], paths['obj']))
+	if os.path.isfile(os.path.join(paths['dir'], paths['obj'])):
+		os.remove(os.path.join(paths['dir'], paths['obj']))
+		
 	# remove data files from user directory
 	for dfile in data_files:
 		if os.path.isfile(dfile.user_copy):
